requirements.txt
googletrans==3.1.0a0
gTTS
requests
gnews
django-allauth
django
newspaper3k